Adds social accounts controller and moves filters to new folder
